refactor: sctructure of project changed, added app.js file
