+++ src/app/file-operations/create.js
@@ -3,7 +3,7 @@ import fs from 'fs';
 export function create(filePath) {
     fs.writeFile(filePath, '', 'utf-8', (err) => {
         if (err) {
-            console.error(`Error creating file: ${err.message}`);
+            throw new Error(`Error creating file: ${error.message}`);
         } else {
             console.log(`File created: ${filePath}`);
         }
