+++ src/app/file-operations/delete.js
@@ -3,7 +3,7 @@ import fs from 'fs';
 export function deleteFile(filePath) {
     fs.unlink(filePath, (err) => {
         if (err) {
-            console.error(`Error deleting file: ${err.message}`);
+            throw new Error(`Error deleting file: ${error.message}`);
         } else {
             console.log(`File deleted: ${filePath}`);
         }
