+++ src/app/file-operations/rename.js
@@ -3,7 +3,7 @@ import fs from 'fs';
 export function rename(oldPath, newPath) {
     fs.rename(oldPath, newPath, (err) => {
         if (err) {
-            console.error(`Error renaming file:`, err);
+            throw new Error(`Error renaming file: ${error.message}`);
         } else {
             console.log(`${oldPath} has been renamed to ${newPath}.`);
         }
