+++ src/app/os-info/eol.js
@@ -1,9 +1,5 @@
 import os from 'os';
 
 export function printEOL() {
-    try {
-        console.log(JSON.stringify(os.EOL));
-    } catch (error) {
-        console.error(`Error: ${error.message}`);
-    }
+    console.log(JSON.stringify(os.EOL));
 }
